Make Krita build with KOXML_USE_QDOM defined
This "fixes" all the places where KoXml::asQDomElement was used,
which was broken. It looks like nobody built like this since, oh,
2010 or so... It's a first step towards not using the KoXmlReader
code at all anymore.

There is no visible performance regression, but I didn't do real
benchmarking. But at least, with the ifdef working, we can do
benchmarking again.

In the text shape, some parts are just rudely ifdeffed out, mostly
to do with RDF, which krita doesn't need anyway, or changetracking,
which krita also doesn't need.

The really tricky bit is that in several places we used KoXmlDocument
with stripSpaces set to true, and I'm not sure what not being able
to set that flag will have.

KOXML_USE_QDOM is defined now, so Wolthera can continue with
the calligraphy tool; before merging to master, we should undefine
it again. Maybe we should make it a cmake option, if we're not just
going to remove KoXmlReader altogether.

CCMAIL:kimageshop@kde.org

CC'ing the calligra developers since this might be of interest for
Calligra, too: either calligra should remove KOXmlReaderForward.h,
or make the ifdef work again -- which is likely much tricker than
it was for krita.

CCMAIL:calligra-devel@kde.org