BUG:376098 Warn the user when saving a read-only file
And use QSaveFile to be smart about the permissions and ownership.
Note that this demands a very filthy hack for saving using KoStore:
we need to make sure KArchive cannot close() the iodevice, because
that aborts if it's a QSaveFile.

For other file types, like tiff and exr, we cannot save to io
devices, we can only pass filenames to the relevant libraries. For
those, add a flag to KisImportExportFilter that IO devices aren't
supported, but rather ignored.