Don't let the user close the document while saving is in progress
Ideally, we should postpone closing till the saving is finished,
but it is impossible to do right now, because doc->setModified()
is called outside the saving lock held. Later, when we refactor
the saving code and the modified state will be cleared under
the lock, we can easily use KisSignalCompressor to close the
document.

The crash itself happened because the hack in KisPart::removeView()
just skipped the deinitialization of the view and the input manager
was still connected to the canvas that was destroyed.

BUG:378538