Don't abort if the tile directory has the wrong size
This allows us to at least load the file referenced in

BUG:379889

which is incomaptible with xcf version v0003, but still carries
that version number. Now it loads in krita... But no guarantees
about 2.9 files. Mostly, they will still crash krita because
the xcftools library aborts hard on encountering a problem, which
we cannot prevent.