Implement saving of SVG text!
Now the biggest part of font properties is loaded/saved using
KoSvgTextProperties object. That allows us not to write the property
that will be inherited from the parent object.

This patch also fixes some whitespace loading problems. The situation
with whitespaces is still not ideal, but it works much better than
before.