FEATURE: Add a docker that gives access to svg symbol libraries
You can drag & drop svg items from the docker to the canvas. This
uses the new KoDrag and KoSvgPaste support: the KoSvgPaste methods
are made static for easier access.

The icons are still a bit iffy, and there is no support for editing
collections, searching collections or tagging collections yet, one
needs inkscape for that.

The objects are dropped as groups of shapes, instead of symbols,
and that is intentional: this is meant for things like speech bubbles
which more often need editing than being exactly the same in the whole
document. Krita's still not map-making software...

CCMAIL:kimageshop@kde.org