Implemented predefined transformations for shapes
This patch implements actions in the shape's context menu:

1) Rotate Shape +-90deg, 180deg
2) Mirror X,Y
3) Reset all applied transformations (was not
   available previously)

This patch also refactors KoSelection a bit. The idea
with tracking shape's matrices and check for their
consistency, in the end, was incorrect. The shapes change
their transformations one-by-one, not atomically. That
makes all the iterative calls checkMatricesConsistent()
return false.

Now the strategies that cause visible transformations
to the selection outline just include the selection
into the list of transformed shapes and transform it
manually.

CC:kimageshop@kde.org
Ref T1005