Change the way how paintops are saved.
This does two things:

1. When overwriting a paintop, the unnamed file is the last file, and the new file is a backup of the old file.
2. The backup are named name-date-time.kpp, with the ISO standard for timestampes: big to small.

This makes it easier to manage brush creation, because it removes the need to rename the last file before adding it to a bundle. It also makes managing backups easier, as you do not need to remove the file from the blacklist, only removing the timestap. Finally, timestamps are easier to read.

CCMAIL: kimageshop@kde.org
BUG:330251
Differential Revision: https://phabricator.kde.org/D6314
Reviewed by rempt, scottyp, dmitryK, timotheegiet and wysteria[m] on IRC, thanks for testing!