Initialization of drawing angle for Polygon and Polyline tools
Summary:
This revision changes the behavior of the Polygon and Polyline tools when using a brush that is dependent on drawing angle. Without the patch, the drawing angle is not initialized at the start of a polyline, resulting in a first dab that always behaves as if the angle is zero. With the patch, the drawing angle is initialized based on the direction from the first point to the second. (The Line tool already behaves this way before the patch.)

{F3797297}

Test Plan:
The difference can be seen when using an angle-dependent brush, such as the Block_basic preset. To test the change, try drawing a vertical line with the Polyline or Polygon tool. Without the patch, the resulting polyline will have an initial dab that faces directly to the right (zero rotation). With the patch, the initial dab points in the direction of the first line segment.

The revision causes no new unit test failures on the author's machine.

Reviewers: #krita, dkazakov

Reviewed By: #krita, dkazakov

Subscribers: dkazakov, allenmarshall

Differential Revision: https://phabricator.kde.org/D6448