Disable freehand stroke timeout when airbrushing
Summary: This revision disables the freehand stroke timeout used by Basic and Weighted smoothing when airbrushing is enabled. This allows airbrushing with a low rate to work with Basic smoothing.

Test Plan:
The difference in behavior can be seen as follows.
1. Set up a brush with the Airbrush option enabled and the rate set to a low value (e.g. 1).
2. Select the freehand tool and enable Basic smoothing.
3. Start a freehand stroke, then hold the cursor still without generating any pointer events (easiest to do with a mouse).
4. With the patch, the airbrush feature should cause dabs to be painted slowly according to the rate configured in step 1. Without the patch, the airbrushing will not happen, because the airbrush rate is slower than the stroke timeout used by Basic smoothing.

The revision causes no new unit test failures on the author's machine.

Reviewers: #krita, rempt

Reviewed By: #krita, rempt

Subscribers: woltherav, allenmarshall

Tags: #krita

Differential Revision: https://phabricator.kde.org/D6463