Fix progress reporting for unthreaded tools like Gradient Tool
The patch just creates an intermediate KoProgressUpdater in
KisViewManager that proxies the progress update requests and
calls processEvents() aggordingly.

WARNING: this is not a regression, but if one presses Ctrl+S while
gradient is painting something, there will be a light deadlock. The
progress of the gradient will stop until the user closes the waiting
window. As soon the barrier waiting window is closed, gradient action
is continued.

I don't think this bug can be fixed without rewriting the gradient
to strokes system.