Fix adding the currently saved file to the list of recent files
The current implementation is a bit weird. It first asks all the
windows to add the file to the list, then asks all the windows to save
these files to a config file. That causes really weird effects, like
the config reread incorrectly after such save.

One day it will need to be refactored...