Work on unit configuration saving when dialog are closed for the image size plugin.
Summary:
The most noticeable novelty is that unit selections now maintain when a dialog is closed and re-opened.
Parameters are saved in the config files, so even when krita is closed they are not lost.

A few improvements have been done on the spinboxunitmanager, allowing to:
 - sync them (so that it is possible to have the same unit in both, but different references for relative sizes).
 - Use % as length unit in spin box. The document aware spinbox is able to see if it is equivalent to a vw or vh.

To test the changes, just create a new image, open any dialog like canva size, image size, layer size or the offset image dialog.

Ref T1621

Reviewers: dkazakov, rempt

Maniphest Tasks: T1621

Differential Revision: https://phabricator.kde.org/D6592