Initialze gmic_qt_get_cropped_images handler's cropRect to a unit square
prepareCroppedImages() expects normalized coordinates to define the area of an image which will be sent to G'MIC.
These coordinates are normally retrieved from the message passed from G'MIC to Krita, but in the case of an error, the dimensions of the image were used instead.
Since the image dimensions are not normalized, this was incorrect.

This commit causes a unit square to be used when G'MIC didn't send valid coordinates, thus using the full area of the canvas or selection.