Another attempt at making the gradients work again.
This time, we store an XML that can be read by the stop gradient object. This does not retroactively fix older files which had issues with the old method.
However, if the old file did not have an issue with the old method, it will be loaded properly and saved with the new method.

BUG:375720
This should be possible to backport.