Add ComponentsOrdered() and add two scripts for exporting to GPL and SVG
ComponentsOrdered gives the components ordered to the displayorder, which can be different from the regular order.
GPL is of course used by Gimp, Krita, Inkscape, MyPaint and many others.
The SVG can be opened by Inkscape and understood as swatches, however is incomplete.
It still needs colorprofile definitions, but those in turn need the location of the profile.

Both would also benefit from simple GUIs and being turned into an extension or something.
Don't forget to remove the sipkritapart0 file before building if you want to use these.

Ref T112