Add a function to get the total colors from a palette, and add stuff to palette docker.
The palette docker now has a dropdown box that can be used to search for colors in a palette,
which works with partial searches and shows icons for the swatch that are different depending
on whether it is a spot color or not.

Also simple buttons that activate add entry, add group and remove group. These will need icons but
having them in their most basic form will be good for now. Outside of a missing function to get the
current foreground closest color selected and no proper resource selector for the palettes outside of
a simple combobox, the python palette docker is now up to date with the regular palette docker in terms
of functionality.

Uhm, yeah, this requires a pysippart removal again, sorry.

Ref T112