Add a Quick Settings Python Docker.
This docker allows people to select Size, Opacity and Flow from a list of useful presets, similar to dockers found in other painting programs.
As it is written in Python, people can disable it easily or even modify it to their tastes, to for example,
to change the list of presets.

BUG:378942