Implement common code for rendering frame sequences
Right now we have a pair of common classes that can be reused for
rendering a sequence of frames with user feedback, multithreading and
cancellation capabilities.

KisAsyncAnimationRendererBase is a base class for rendering a single
frame in a separate thread using a dedicated image. The user should just
reimplement two pure virtual callback functions to make it save real data
on frame completion and error.

KisAsyncAnimationRenderDialogBase is a class that constructs a set of
renderer objects, multithreads them and shows the progress reports to
the user (if needed).