Optimize KisUpdaterContext to process multiple jobs in bunches
The patch fixes a problem of waiting/sleepy threads when the number
of cores is big anough and when the size of the jobs is small enough.

Basically, we shouldn't pass the job back to a thread pool if there
are any other jobs pending. Passing a job to a pool will mean getting
the thread into sleep, which we should avoid. Therefore we group the
jobs into batches so they can be executed without any interruption.

See a comment in kis_update_job_item.h for more info.