Change texture formats used with OpenGL ES 3
This removes the use of certain formats that could crash when OpenEXR is
not present. The modified code also does not work with OpenGL ES 3
implementations without GL_EXT_texture_format_BGRA8888 support with
certain image formats. Instead, a safe assert will trigger and the image
will be rendered with the red and blue channels swapped.

Maniphest Tasks: T6696