Implement a hazardous way of fetching tiles from the hash table
Since we use our own shared pointers for tiles, we can relax the
requirements for their access a bit. Our shared pointer are thread-
safe: when a reader and writer access the pointer concurrently, the
reader is guaranteed to get either a valid pointer or null.

This is the base of a new, quick and dirty way of fetching tiles
called getTileMinefieldWalk()