Add KisLoggingManager for capturing log messages
This class is mainly for capturing log messages from the Qt logging
infrastructure (QLoggingCategory and qDebug etc.) for providing
diagnostics information for use in bug report and showing them in the UI
for convenience. For its limited use at the moment, this class is not
made to be thread-safe in any sense.

It can be extended in the future for more use, though it should be made
thread-safe for any more complex tasks.