Fix the offset in the triangle color selector.
This, combined with the recent speed up should fix two of the big issues with
the internal selector. The other issues are a desyncing between the two selector widgets
due to stupid rounding errors, and whether or not it should be possible to have the hue ring
stuck to one color. These will both take a lot of rewiring.

Anyway, test plz when you have the chance.

CCBUG:372221