BUG:378147 Fix reloading file layers twice.
Basically, this was caused by the backups.
Krita handles backups by saying "ok, you, old file, rename to backup,
and new file, you become the url of old file". This renaming is what
signals QFileSystemWatcher, and because it is a rename, it removes it
from the list of watched files.

So what we do is that when we get our change in, we check if the file
exists and is still being watched, and if so we readd it.