Refactor KisPaintInformation to not keep a link to KisDistanceInformation
We still need to register/unregister the distance information into the
paint information, but the link is not kept. The paint information is
just initialized with the values from the distance information and since
that moment the sensors like drawingAngle() can work independently from
the distance information. It means that one can *copy* paint information
and it will still be fully capable.