Refactor KisDabCache into two smaller classes
It is needed to make it work in multithreaded environment. Now we have
a base class that encapsulates the decision whether to generate a new
dab or use a cached version. And the cache class just stores the cached
version and uses the base class to make a decision about caching.

This base class will be reused in the threadable dabs queue.