Implement KisSharedThreadPoolAdapter concept
Shared thread pool concept is a wrapper around a usual thread pool
that allows the user to wait for a subset of running jobs to complete.

The jobs added to the thread pool via shared thread pool adapter are
treated separately and can be waitForDone() without waiting for all the
jobs of the thread pool to complete.