Only start the reference image loader thread when the docker becomes visible
This won't stop Krita from crashing on startup in a qimageio plugin if
there are broken files in the location where it's looking, but at least
that crash only happens when people use the imagedocker. It should also
help a bit with the load on startup, especially if the Pictures folder
is quite full.

The real issue is, of course, that Qt's qimageio plugin mechanism is
really, really, really dangerous.

CCBUG:384641