Remove the configuration for gmic-qt
There are now two ways of finding gmic-qt: have the executable
built with the same Qt and other libraries as krita next to the
krita executable (should work for both our builds as well as for
distributions that package both krita and gmic-qt), or locate
a folder that starts with gmic next to the krita executable. The
folder can contain the version of gmic-qt that is build by gmic.eu.

For a next iteration, I am considering also looking into
QStandardPaths::AppDataLocation or QStandardPaths::AppLocalDataLocation
to find a gmic-qt plugin folder, but for the current release, this
should work.

CCMAIL:kimageshop@kde.org