Don't remove *. from the suffixes returned by KoMimeDatabase
Following commit

commit bdb2e7fc2d75016ea98f0106ad731d5a02fa1b53
Author: Dmitry Kazakov <dimula73@gmail.com>
Date:   Fri Sep 22 21:39:56 2017 +0300

    KisMimeDatabase should contain suffixes without '*.' prefix

This code should be adapted as well. The problem is that the file
dialogs need *.suffix to filter the files, and return that as the
chosen filter as well. QFileDialog until Qt 5.9 didn't have
a QString QFileDialog::selectedMimeTypeFilter() const method...

If we make 5.9 our minimum version, we can rewrite KoFileDialog to
be logical and consistent.