Rewrite massive bltFixed(rect, dabs) implementation
tldr: new bltFixed(rect, dabs) is up to twice faster than the old
      one and about 20% faster than conventional dabs rendering with
      calling bltFixed(dab) sequentially.

## Conventional dab rendering method (before mt-work) ##

The paintop just calls bltFixed(dab) sequentially. Each call does
readBytes+compose+writeBytes to a destination paint device. This
algorithm has two drawbacks:

    1) ReadBytes+WriteBytes needs an additional memory allocation and
       takes time

    2) It is very difficult to split this work into multiple threads

## Old dab rendering method (my first attempt) ##

The passed area was split into subareas where the list of overlapping
dabs was the same. Then the dads were rendered over each area. It meant
that each pixel of *dst* device was rendered only *once*.

This method resulted to be almost twice slower than conventional one.
The problem was that the subareas were too small (< 64 pixels), therefore
the tiles were read in a non-sequential way, which was slow.

## New dab rendering method (in this patch) ##

We just render the the dabs sequentally, like we do in conventional
method, but we avoid read/write bytes. Instead we just write directly
to the paint device using iterators. This method is up to 20% faster
than the conventional method when the number of dabs is less than 10,
and has the same speed otherwise. The main benefit of this algorithm
is that it is threadable.