Implement mutated jobs framework
"Mutated jobs" is a special paradigm of dab-jobs in the strokes framework,
that might be added to the stroke's queue right during the execution of
the queue.

Imaging you have a job for generation of a single dab. This job can
create multiple "mutated" jobs that continue the work started by the
original job, but in multiple threads. The original job will somehow
"mutate" into multiple jobs that can be executed in parallel.

The mutated jobs conform the same sequentiality and exclusivity
principles as normal jobs. That is, if you want to wait for completion
of multiple parallel mutated jobs, just add one more sequential (also
mutated) job after them.