[Feature] ASC-CDL color balance filter(Slope, Offset, Power)
This is a simple color balance filter as decided by the American
Society of Cinematographers. It's purpose is to make communicating color
balance easier over different applications.

It is also interesting in that it is so simple it does not clip high-bit depth
values.

There's still some issues with the filter, but they are mostly GUI issues,
so I am pushing this to prevent bitrot. I have noted the possible list of
improvements in the header of the widget class.

BUG:355747
Differential revision: https://phabricator.kde.org/D8184
CCMAIL:kimageshop@kde.org