Fix using the wheel when the palettte docker is very small
This patch solves an inconsistency in the UI where the user is unable
to scrollwheel-zoom-in color swatches in the palette docker after the
same has been reduced to its minimum width, which causes color swatch
size to be set to 8px, a size that cannot be normally set with the
ctrl+scrollwheel shortcut. This color swatch size is now made accessible
with the scroll wheel, and scroll events are now blocked only when trying
to further zoom out.

BUG:385978
Patch by Neviril <nevineviril@yahoo.com> Thanks!