Delay purging of KisTextureTileInfoPoolSingleSize
Adds a KisSignalCompressor to the purge memory logic in KisTextureTileInfoPool, such that the pool items are not freed immediately as the pool allocation reaches 0, but only after a delay of 1s. This fixes a performance hit on some systems, where the pool gets completely freed and reallocated several times a second.

Differential Revision: https://phabricator.kde.org/D8279