Implement a manager for tiles extents to avoid long mutex locks
This manager avoids recalculation of data manager extents
on every dm->clear(). It also resolves thread-safely problems
of tile updates in getTile().

The main problem of this manager is that it doesn't improve any
performance. After we removed the locks from dm->clear(), all the
waits just migrated to the lock of KisTileHashTable... That is
basically the same situation like in "city traffic jams" problem.
This patch moves a jam to a different place.

It the future, we should resolve the jam in the hash tile
(make it lockfree?), then, coupled with this patch, the waiting
problem will be resolved.

With this patch, 79% of the waits are concentrated in the
hash table. And 9% of the waits happen in the scheduler's
queues.