[FEATURE] Add "convert to file layer" command to file layers.
This adds an action that can be triggered on any non referencing layer(so not file or clone layers).

This action will request a file name, and then save the layer(if a group layer, with everything in it)
into a seperate file. It will then replace the layer in the old with with a file layer referencing the
new file.

The idea behind this one is that if people's layer stacks become too complicated,
they can save out a section of it to a file layer, reducing the complexity of the document.

CCMAIL:Kimageshop@kde.org
Differential Revision: https://phabricator.kde.org/D8360