[FEATURE] Add properties dialog to change file layer properties.
This gives the file layers a custom properties dialog to adjust the
location of the referenced file and location of the file layers.

I think this one can be ported, but the other file layer patch cannot because
it uses the clone stuff that all the async saving work brought.

Differential Revision: https://phabricator.kde.org/D8359
BUG:380467
CCMAIL:Kimageshop@kde.org