Fixes two "lost tiles" problems in the multithreaded brush
I spent almost a week on that. The first problem was related to
a misprint in KisSimpleUpdateQueue::addJob() after a bulk-update
refactoring. The second problem was reordering of the updates in
KisCanvasUpdatesCompressor::putUpdateInfo(). Updates should never
be reordered in GUI code, when the data is already fetched from the
image and stored in KisUpdateInfo object.

Ref T7309