Update the color selector dialog on showing it
If the fg or bg color hasn't been set at least once, the color
selector dialog would use the dummy internal KoColor and its
colorspace, which is rgb16. Then, on changing the color using
the spinboxes, a the current colorspace would change if the
document wasn't rgb16; the spinbox would be deleted and Krita
would crash.

This change initializes the dialog to the correct colorspace on
showing it, and moves the deletion of the spinboxes to the event
queue, using deleteLater.

BUG:385864
BUG:386066