Implement a simple cleanup phase for removing non-border-touching areas
In other words, we search for the smallest planar area that does not
touch enough contour pixels. If such area is found, we simply remove it
and fill its pixels with the color of surrounding regions. We repeat
these steps until no such areas left.

NOTE:
At the moment the algorithm fills up conflicting areas only within the
original level of terrain. The next step will be to implement an ability
to crawl up the hills to avoid extra colored contours around the lines.