First version of the "fuzzy" watershed algorithm
The difference with normal watershed is that there is a special "cleanup"
phase that lets the user to paint a bit outside the contours of the
object. These extra marks will be removed if the length of their "foreign"
(different color or height) edge is more than 35% of the total edge length.

The value is going to be configurable, but not atm.