Add explicit metadata and author checkboxes to the png converter.
This adds two checkboxes that need to be explicitely set to store author
and meta data in the png, to ensure the user is fully aware that this will happen.

The author and meta data will not be stored if this is not possible, but lets not attempt
to be too smart and also disable the checkboxes. The tooltip should be sufficient.

Ref T6627
CCBUG:379730