Implement more strict brush update time calculation algorithm
If the brush is fully cached while painting (e.g. "Deevad 2e big brush"),
then dabRenderignTime becomes almost null. And a lot of cached dabs are put
into the queue to be rendered on screen, which might take more than 100ms.

Therefore we should limit the number of dabs that can be fetched from
the queue in one go so that they would fit into the update period.

Ref T7309