Scripter: recognise if script is modified.
This gets around the weird issue where if you load a script and make changes, the run action would only
try to run the original script as it is in the file, and not your changes. Adding a modified function
allows us to run the changes and remind the user that they haven't saved their script.