Invert the filtered device to make it more pleasantly looking
Please note that we do it not in extremely efficient way. The worker
still works with the inverted version, so we should prepare the heightmap
for it every time.

We should refactor the worker later when the algorithm kind of stabilizes.
The problem is that "positive" and "negative" metrics depend on the
heightmap direction atm, so changing the heightmap format may change
the semantics of some functions in the worker.