Implement basic generation of XML from a KoSvgTextShape
This re-uses the KoSvgTextProperties::convertToSvgTexgtAttributes
to generate the CSS style string for each span/p element. The
text is converted to a single <p> element with the lines separated
by <br>; currently only vertical movement is interpreted as newlines,
when we support vertical text, horizontal movement must also be
checked.