Update scripter to allow running of scripts without a main() function
Summary:
Allow scripter to play main()-less scripts without complaining

If a script is typed into the scripter window interactively and saved, scripter keeps a record of the file and automatically loads the script on the next run. When you press play on this auto-loaded script the runner module tries to load the module, then run it's main() method. However, an interactive script may not have a main method.  Scripter should run main() if it's present, but fail silently if it is not present.

Test Plan:
Not sure what to put here.
To trigger the initial bug Run krita, run scripter, create a script without main, save the script then press the play button.

Reviewers: #krita, woltherav

Reviewed By: #krita, woltherav

Subscribers: alvinhochun, rempt, eliakinalmeida, woltherav

Tags: #krita

Differential Revision: https://phabricator.kde.org/D9317