Prevent dataloss when switching between text and svg source.
This patch does two things:

first, the converter when going from html to svg doesn't contain svg, so it comments out the converter.toSVG(svg, styles) function(which would lead to empty strings).

This should proly be replaced with the converter actually creating text-chunks or something from the svg, but eh.

The second bit is does is to prevent toHTML attempting to write style attributes in the br tag? It's a bit odd but it happens without an extra span. Might proly also benefit from reworking.

Differential Revision: https://phabricator.kde.org/D9492