Add a specialised font combobox combo to the text editor.
This is a widget with two qcomboboxes, where one is for the styles and the other for the families. Having styles seperate allows us to remove a lot of fonts from the families, making it far less intimidating a list.

This also filters superfluous fonts, like the helper fonts that come with latex, and there is a possibility to have it use a list of writing systems as well.

Differential Revision: https://phabricator.kde.org/D9320