2nd try of fixing color fluctuations in colorsmudge-dulling brushes
Now the colorsmudge brush does all the blending in the high bit-depth
colorspace. The patch introduces a special wrapper
(KisPrecisePaintDeviceWrapper) that ensures a proper color space is used
when needed. See the apidox in KisPrecisePaintDeviceWrapper.h.

NOTE: this patch fixes **both** dulling and smearing modes!

BUG:348267

Test Plan:
Needs to be tested:
1) Colorsmudge brushes in Dulling mode with Color Rate option enabled
2) Smudge Radius enabled or disabled: both should work correctly
3) Opacity option should work correctly
4) Overlay mode?

Reviewers: #krita
Differential Revision: https://phabricator.kde.org/D9686