Fix selection of the multiline text with a mouse click
After the SVG refactoring there is no more guarantee that the top
left corner of the shape is always placed at 0,0 point of the local
coordinate system. It might be places anywhere. Please consult with
outline() and outlineRect() methods for retrieving the bounding
rect of the shape in a local coordinate system.